refactor(auth): getMe API, 타입, 쿼리 훅을 shared에서 FSD 구조에 맞게 분리 및 정리
