+++ src/entities/auth/auth.api.ts
@@ -1,6 +1,7 @@
 import { api } from "@/shared/lib/axios";
 import { API_BASE_URL } from "@/shared/api/constants";
 import {
+  GetMeResponseDto,
   PostLoginRequestDto,
   PostLoginResponseDto,
   PostLogoutResponseDto,
@@ -9,6 +10,7 @@ import {
 const AUTH_API_URL = {
   LOGIN: `${API_BASE_URL}/api/auth/login`,
   LOGOUT: `${API_BASE_URL}/api/auth/logout`,
+  ME: `${API_BASE_URL}/api/auth/me`,
 };
 
 export const postLogin = async (
@@ -20,5 +22,10 @@ export const postLogin = async (
 
 export const postLogout = async (): Promise<PostLogoutResponseDto> => {
   const { data } = await api.post(AUTH_API_URL.LOGOUT);
+  return data;
+};
+
+export const getMe = async (): Promise<GetMeResponseDto> => {
+  const { data } = await api.get(AUTH_API_URL.ME);
   return data;
 };
