+++ src/entities/auth/auth.types.ts
@@ -1,3 +1,5 @@
+import { type JWTPayload } from "jose";
+
 export type PostLoginRequestDto = {
   email: string;
   password: string;
@@ -9,4 +11,9 @@ export type PostLoginResponseDto = {
 
 export type PostLogoutResponseDto = {
   success: boolean;
+};
+
+export type GetMeResponseDto = {
+  isLoggedIn: boolean;
+  user?: JWTPayload;
 };
